# SpanHelpers IndexOfAny-methods use sentinel-value