Instead of having _valueLength_-times an if-check in the loop, a sentinel value is used and hence only one final checks needs to done.

Perf-win depends on the size of _valueLength_, the greater the more win.