@jkotas 2017-12-17
Do you have any numbers to demonstrate the perf-win?