@gfoidl 2017-12-17
@jkotas I'm just running better benchmarks and will post results here.