@begay 2017-12-17
To fast I m just learning  simplify language

On Dec 17, 2017 9:40 AM, "Günther Foidl" <notifications@github.com> wrote:

@jkotas <https://github.com/jkotas> I'm just running better benchmarks and
will post results here.

—
You are receiving this because you are subscribed to this thread.
Reply to this email directly, view it on GitHub
<https://github.com/dotnet/corefx/pull/25970#issuecomment-352268456>, or mute
the thread
<https://github.com/notifications/unsubscribe-auth/Af8pHyq1jq1svEXMu8vRyKJZaLCdjZsKks5tBUP_gaJpZM4REtAS>
.