@begay 2017-12-17
I don't understand how to write to you or understand the language I'm new
still can you do it in layman's terms

On Dec 17, 2017 9:55 AM, "Jan Kotas" <notifications@github.com> wrote:

> *@jkotas* commented on this pull request.
> ------------------------------
>
> In src/System.Memory/src/System/SpanHelpers.byte.cs
> <https://github.com/dotnet/corefx/pull/25970#discussion_r157373729>:
>
> > @@ -80,16 +77,13 @@ public static int LastIndexOfAny(ref byte searchSpace, int searchSpaceLength, re
>              if (valueLength == 0)
>                  return 0;  // A zero-length sequence is always treated as "found" at the start of the search space.
>
> -            int index = -1;
> +            int index = int.MinValue;
>
> You may be able to just use int index = -1; here, and then you do not
> need to normalize it at the end.
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/dotnet/corefx/pull/25970#pullrequestreview-84005544>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/Af8pHwkYAI7mQDvaaaX7BtaU6WmMGrTSks5tBUdjgaJpZM4REtAS>
> .
>