@gfoidl 2017-12-17
# Results

Code for benchmarks: https://gist.github.com/gfoidl/842f3ca4a80e532515f4476cd92cae66

## Span<byte>.IndexOfAny

``` ini

BenchmarkDotNet=v0.10.11, OS=Windows 7 SP1 (6.1.7601.0)
Processor=Intel Core i7-3610QM CPU 2.30GHz (Ivy Bridge), ProcessorCount=8
Frequency=2241054 Hz, Resolution=446.2186 ns, Timer=TSC
.NET Core SDK=2.1.2
  [Host]     : .NET Core 2.0.3 (Framework 4.6.25815.02), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.3 (Framework 4.6.25815.02), 64bit RyuJIT


```
|        Method | BufferSize | ValuesSize |       Mean |     Error |    StdDev | Scaled |
|-------------- |----------- |----------- |-----------:|----------:|----------:|-------:|
|    **IndexOfAny** |       **1024** |          **4** |   **429.3 ns** | **0.4782 ns** | **0.4239 ns** |   **1.00** |
| IndexOfAnyNew |       1024 |          4 |   394.9 ns | 0.4722 ns | 0.3687 ns |   0.92 |
|    **IndexOfAny** |       **1024** |         **10** | **1,033.3 ns** | **1.7612 ns** | **1.5613 ns** |   **1.00** |
| IndexOfAnyNew |       1024 |         10 |   950.1 ns | 2.4577 ns | 1.9188 ns |   0.92 |
|    **IndexOfAny** |       **4096** |          **4** | **1,181.2 ns** | **0.8200 ns** | **0.7670 ns** |   **1.00** |
| IndexOfAnyNew |       4096 |          4 | 1,147.8 ns | 0.8821 ns | 0.8252 ns |   0.97 |
|    **IndexOfAny** |       **4096** |         **10** | **2,924.2 ns** | **8.1831 ns** | **6.3888 ns** |   **1.00** |
| IndexOfAnyNew |       4096 |         10 | 2,828.9 ns | 2.2138 ns | 2.0707 ns |   0.97 |

## Span<byte>.LastIndexOfAny

Here I can't measure any notable difference. When running the benchmark multiple times once IndexOfAny is slightly faster, and once the method from this PR is faster.  
For me this seems strange, because the JITed code is shorter and more precise in the new variant, so it "must" be faster. Maybe my code for the benchmark is not ideal to reflect this exactly.

### Old JITed code

Note: method got inlined
```asm
000007fe`773adb60 System.SpanHelpers.LastIndexOfAny(Byte ByRef, Int32, Byte ByRef, Int32)
			int index = -1;
   ^^^^^^^^^^^^^^^
000007fe`773adb77 41beffffffff    mov     r14d,0FFFFFFFFh
			for (int i = 0; i < valueLength; i++)
        ^^^^^^^^^
000007fe`773adb7d 4533ff          xor     r15d,r15d
			for (int i = 0; i < valueLength; i++)
                   ^^^^^^^^^^^^^^^
000007fe`773adb80 85f6            test    esi,esi
000007fe`773adb82 7e2f            jle     000007fe`773adbb3
000007fe`773adb84 4963d7          movsxd  rdx,r15d
000007fe`773adb87 0fb61413        movzx   edx,byte ptr [rbx+rdx]
000007fe`773adb8b 488bcf          mov     rcx,rdi
000007fe`773adb8e 448bc5          mov     r8d,ebp
000007fe`773adb91 e802f3ffff      call    000007fe`773ace98
				if (tempIndex != -1)
    ^^^^^^^^^^^^^^^^^^^^
000007fe`773adb96 83f8ff          cmp     eax,0FFFFFFFFh
000007fe`773adb99 7410            je      000007fe`773adbab
					index = (index == -1 || index < tempIndex) ? tempIndex : index;
     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
000007fe`773adb9b 4183feff        cmp     r14d,0FFFFFFFFh
000007fe`773adb9f 7407            je      000007fe`773adba8
000007fe`773adba1 443bf0          cmp     r14d,eax
000007fe`773adba4 7c02            jl      000007fe`773adba8
			for (int i = 0; i < valueLength; i++)
                                    ^^^
000007fe`773adba6 eb03            jmp     000007fe`773adbab
000007fe`773adba8 448bf0          mov     r14d,eax
000007fe`773adbab 41ffc7          inc     r15d
000007fe`773adbae 443bfe          cmp     r15d,esi
000007fe`773adbb1 7cd1            jl      000007fe`773adb84
			return index;
   ^^^^^^^^^^^^^
000007fe`773adbb3 418bc6          mov     eax,r14d
```

### New JITed code

Note: method got inlined
```asm
000007fe`773adb60 System.SpanHelpers.LastIndexOfAnyNew(Byte ByRef, Int32, Byte ByRef, Int32)
			int index = -1;
   ^^^^^^^^^^^^^^^
000007fe`773adb77 41beffffffff    mov     r14d,0FFFFFFFFh
			for (int i = 0; i < valueLength; i++)
        ^^^^^^^^^
000007fe`773adb7d 4533ff          xor     r15d,r15d
			for (int i = 0; i < valueLength; i++)
                   ^^^^^^^^^^^^^^^
000007fe`773adb80 85f6            test    esi,esi
000007fe`773adb82 7e22            jle     000007fe`773adba6
000007fe`773adb84 4963d7          movsxd  rdx,r15d
000007fe`773adb87 0fb61413        movzx   edx,byte ptr [rbx+rdx]
000007fe`773adb8b 488bcf          mov     rcx,rdi
000007fe`773adb8e 448bc5          mov     r8d,ebp
000007fe`773adb91 e802f3ffff      call    000007fe`773ace98
				if (tempIndex > index) index = tempIndex;
    ^^^^^^^^^^^^^^^^^^^^^^
000007fe`773adb96 413bc6          cmp     eax,r14d
000007fe`773adb99 7e03            jle     000007fe`773adb9e
				if (tempIndex > index) index = tempIndex;
                           ^^^^^^^^^^^^^^^^^^
000007fe`773adb9b 448bf0          mov     r14d,eax
			for (int i = 0; i < valueLength; i++)
                                    ^^^
000007fe`773adb9e 41ffc7          inc     r15d
000007fe`773adba1 443bfe          cmp     r15d,esi
000007fe`773adba4 7cde            jl      000007fe`773adb84
			return index;
   ^^^^^^^^^^^^^
000007fe`773adba6 418bc6          mov     eax,r14d
```